requirements.txt
flask
numpy
pandas
requests
google-generativeai
scikit_learn==1.5.1
gunicorn>=20.1.0
joblib
python-dotenv
